+++ packages/wallet-connectors/src/WalletConnection.ts
@@ -35,11 +35,6 @@ export interface WalletConnection {
     ping(): Promise<void>;
 
     /**
-     * @return The account that the wallet currently associates with this connection.
-     */
-    getConnectedAccount(): Promise<string | undefined>;
-
-    /**
      * Returns a JSON-RPC client that is ready to perform requests against some Concordium Node connected to network/chain
      * that the connected account lives on.
      *
@@ -177,8 +172,9 @@ export interface WalletConnectionDelegate {
     /**
      * Notification that the given {@link WalletConnection} has been established.
      * @param connection Affected connection.
+     * @param address The address of the initially connected account.
      */
-    onConnected(connection: WalletConnection): void;
+    onConnected(connection: WalletConnection, address: string | undefined): void;
 
     /**
      * Notification that the given {@link WalletConnection} has been disconnected.